// src/api-client/models/donation-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Lifecycle state.
 * @export
 * @enum {string}
 */

export const DonationStatus = {
    Draft: 'DRAFT',
    Offered: 'OFFERED',
    Accepted: 'ACCEPTED',
    Declined: 'DECLINED',
    ReadyForPickup: 'READY_FOR_PICKUP',
    DriverEnRoute: 'DRIVER_EN_ROUTE',
    Delivered: 'DELIVERED',
    Cancelled: 'CANCELLED'
} as const;

export type DonationStatus = typeof DonationStatus[keyof typeof DonationStatus];